+++ src/utils/networks/types/AA.js
@@ -2,12 +2,12 @@ import aa from '@/assets/images/networks/aa.svg';
 
 export default {
   name: 'AA',
-  name_long: 'Arthera',
+  name_long: 'Arthera Mainnet',
   homePage: 'https://arthera.net/',
   blockExplorer: 'Blockscout',
-  blockExplorerTX: 'https://explorer-test.arthera.net/tx/[[txHash]]',
-  blockExplorerAddr: 'https://explorer-test.arthera.net/address/[[address]]',
-  chainID: 10243,
+  blockExplorerTX: 'https://explorer.arthera.net/tx/[[txHash]]',
+  blockExplorerAddr: 'https://explorer.arthera.net/address/[[address]]',
+  chainID: 10242,
   tokens: import('@/_generated/tokens/tokens-goerli.json').then(
     val => val.default
   ),
@@ -16,7 +16,7 @@ export default {
   ),
   icon: aa,
   currencyName: 'AA',
-  isTestNetwork: true,
+  isTestNetwork: false,
   isEthVMSupported: {
     supported: false,
     url: null,